Worldgen Rework [WIP]

+++ src/main/java/gregtech/api/util/VoltageTier.java
@@ -5,22 +5,22 @@ import net.minecraft.util.StringIdentifiable;
 
 public enum VoltageTier implements StringIdentifiable {
 
-    ULV("ultra_low_voltage", 8L, 0xDCDCDC),
-    LV("low_voltage", 32L, 0xDCDCDC),
-    MV("medium_voltage", 128L, 0xFF6400),
-    HV("high_voltage", 512L, 0xFFFF1E),
-    EV("extreme_voltage", 2048L, 0x808080),
-    IV("insane_voltage", 8192L, 0xF0F0F5),
-    LuV("ludicrous_voltage", 32768L, 0xF0F0F5),
+    ULV("ultra_low_voltage", 8, 0xDCDCDC),
+    LV("low_voltage", 32, 0xDCDCDC),
+    MV("medium_voltage", 128, 0xFF6400),
+    HV("high_voltage", 512, 0xFFFF1E),
+    EV("extreme_voltage", 2048, 0x808080),
+    IV("insane_voltage", 8192, 0xF0F0F5),
+    LuV("ludicrous_voltage", 32768, 0xF0F0F5),
     ZPM("zpm_voltage", 131072, 0xF0F0F5),
     UV("ultimate_voltage", 524288, 0xF0F0F5),
     MAX("maximum_voltage", Integer.MAX_VALUE, 0xF0F0F5);
 
     private final String name;
-    private final long voltage;
+    private final int voltage;
     private final int voltageColor;
 
-    VoltageTier(String name, long voltage, int voltageColor) {
+    VoltageTier(String name, int voltage, int voltageColor) {
         this.name = name;
         this.voltage = voltage;
         this.voltageColor = voltageColor;
@@ -34,7 +34,7 @@ public enum VoltageTier implements StringIdentifiable {
         return GTValues.MODID + ".voltage." + this.name;
     }
 
-    public long getVoltage() {
+    public int getVoltage() {
         return voltage;
     }
